Exception Handler - delete

+++ src/main/java/com/example/demo2/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.example.demo2.entities.User;
 import com.example.demo2.repositories.UserRepository;
+import com.example.demo2.services.exceptions.DatabaseException;
 import com.example.demo2.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -32,9 +35,18 @@ public class UserService {
 		return repository.save(obj);
 	}
 	
-	public void delete (Long id) {
-		
-		repository.deleteById(id);
+	public void delete(Long id) {
+		try {
+			if (!repository.existsById(id)) {
+		        System.out.println("Erro: Usuário não encontrado!");
+		        throw new ResourceNotFoundException(id);
+		    }
+			repository.deleteById(id);
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update (Long id, User obj) {
